+++ task_test.go
@@ -2120,3 +2120,49 @@ func TestSilence(t *testing.T) {
 
 	buff.Reset()
 }
+
+func TestForce(t *testing.T) {
+	tests := []struct {
+		name     string
+		env      map[string]string
+		force    bool
+		forceAll bool
+	}{
+		{
+			name:  "force",
+			force: true,
+		},
+		{
+			name:     "force-all",
+			forceAll: true,
+		},
+		{
+			name:  "force with gentle force experiment",
+			force: true,
+			env: map[string]string{
+				"TASK_X_GENTLE_FORCE": "1",
+			},
+		},
+		{
+			name:     "force-all with gentle force experiment",
+			forceAll: true,
+			env: map[string]string{
+				"TASK_X_GENTLE_FORCE": "1",
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var buff bytes.Buffer
+			e := task.Executor{
+				Dir:      "testdata/force",
+				Stdout:   &buff,
+				Stderr:   &buff,
+				Force:    tt.force,
+				ForceAll: tt.forceAll,
+			}
+			require.NoError(t, e.Setup())
+			require.NoError(t, e.Run(context.Background(), taskfile.Call{Task: "task-with-dep", Direct: true}))
+		})
+	}
+}