+++ taskfile/call.go
@@ -5,4 +5,5 @@ type Call struct {
 	Task   string
 	Vars   *Vars
 	Silent bool
+	Direct bool // Was the task called directly or via another task?
 }